% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dwellings_denver}
\alias{dwellings_denver}
\title{Denver residential dwelling sales for 2013}
\format{
A data frame with columns:
\describe{
\item{parcel}{The variable is character. The parcel id}
\item{nbhd}{The variable is numeric. Neigborhood of the home}
\item{abstrprd}{The variable is numeric. No clue}
\item{livearea}{The variable is numeric. Square footage that is liveable}
\item{finbsmnt}{The variable is numeric. Square footage finished in the basement}
\item{basement}{The variable is numeric. Total square footage of the basement}
\item{yrbuilt}{The variable is numeric. Year the home was built}
\item{condition}{The variable is character. Condition of the home (6 levels provided)}
\item{quality}{The variable is character. A letter ratting}
\item{totunits}{The variable is numeric. How many dwelling units in the building}
\item{stories}{The variable is numeric. The number of stories}
\item{gartype}{The variable is character. Details on the garage type}
\item{nocars}{The variable is numeric. size of the garage in cars}
\item{xtraffic}{The variable is logical. emtpy}
\item{floorlvl}{The variable is numeric. Mostly for apartments.  What level the living unit is on}
\item{numbdrm}{The variable is numeric. Number of bedrooms}
\item{numbaths}{The variable is numeric. Number of bathrooms}
\item{arcstyle}{The variable is character. Type of home}
\item{sprice}{The variable is numeric. Selling price}
\item{deduct}{The variable is numeric. Deduction from the selling price}
\item{netprice}{The variable is numeric. Net price of home}
\item{tasp}{The variable is numeric. Tax assesed selling price}
\item{smonth}{The variable is numeric. Month sold}
\item{syear}{The variable is numeric. Year sold}
\item{qualified}{The variable is character. Q or U with 66 percent Q}
\item{status}{The variable is character. I or V with over 90 percent I}
}
}
\source{
\url{https://www.denvergov.org/opendata/dataset/city-and-county-of-denver-real-property-sales-book-2013}
}
\usage{
dwellings_denver
}
\description{
Attributes of each dwelling with their selling price for homes that sold in Denver in 2013
}
\examples{
\dontrun{
dwellings_denver
}
}
\keyword{datasets}
