% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dwellings_ml}
\alias{dwellings_ml}
\title{Denver residential dwelling sales for 2013}
\format{
A data frame with columns:
\describe{
\item{parcel}{The variable is character. The parcel id}
\item{abstrprd}{The variable is numeric. No clue}
\item{livearea}{The variable is numeric. Square footage that is liveable}
\item{finbsmnt}{The variable is numeric. Square footage finished in the basement}
\item{basement}{The variable is numeric. Total square footage of the basement}
\item{yrbuilt}{The variable is numeric. Year the home was built}
\item{totunits}{The variable is numeric. How many dwelling units in the building}
\item{stories}{The variable is numeric. The number of stories}
\item{nocars}{The variable is numeric. size of the garage in cars}
\item{numbdrm}{The variable is numeric. Number of bedrooms}
\item{numbaths}{The variable is numeric. Number of bathrooms}
\item{sprice}{The variable is numeric. Selling price}
\item{deduct}{The variable is numeric. Deduction from the selling price}
\item{netprice}{The variable is numeric. Net price of home}
\item{tasp}{The variable is numeric. Tax assesed selling price}
\item{smonth}{The variable is numeric. Month sold}
\item{syear}{The variable is numeric. Year sold}
\item{condition_AVG}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{condition_Excel}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{condition_Fair}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{condition_Good}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{condition_VGood}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{quality_A}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{quality_B}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{quality_C}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{quality_D}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{quality_X}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{gartype_Att}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{gartype_Att/Det}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{gartype_CP}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{gartype_Det}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{gartype_None}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{gartype_att/CP}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{gartype_det/CP}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_BI-LEVEL}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_CONVERSIONS}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_END UNIT}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_MIDDLE UNIT}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_ONE AND HALF-STORY}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_ONE-STORY}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_SPLIT LEVEL}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_THREE-STORY}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_TRI-LEVEL}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_TRI-LEVEL WITH BASEMENT}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_TWO AND HALF-STORY}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{arcstyle_TWO-STORY}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{qualified_Q}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{qualified_U}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{status_I}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{status_V}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
\item{before1980}{The variable is numeric. 0 or 1 to mark building having attriubte as 1}
}
}
\source{
\url{https://www.denvergov.org/opendata/dataset/city-and-county-of-denver-real-property-sales-book-2013}
}
\usage{
dwellings_ml
}
\description{
Attributes of each dwelling with their selling price in machine learning format
}
\examples{
\dontrun{
dwellings_ml
}
}
\keyword{datasets}
